requirements/base.txt
psycopg==3.0.16
psycopg2==2.9.3
Django==3.2
pytz==2022.1
dj-database-url==1.0.0
psycopg2-binary==2.9.3
django-decouple==2.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1

django-allauth==0.51.0
django-ckeditor==6.4.2
django-ckeditor-5==0.1.6
django-crispy-forms==1.14.0
django-js-asset==2.0.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0

pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0

requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
stripe==3.5.0
urllib3==1.26.11
